Add Options Dialog Fix bug recent activity list (crash items are less of 30) Fix bad commit in AddFeedDialog
git-svn-id: svn://svn.code.sf.net/p/tvudownloader/code@42 4e9d9500-3722-a143-8048-b556883094bb

+++ tvu/Hostory.cs
@@ -206,8 +206,10 @@ namespace tvu
 
             }
 
-            myFileHistoryList.RemoveRange(30, myFileHistoryList.Count - 30);
-
+            if (myFileHistoryList.Count > 30)
+            {
+                myFileHistoryList.RemoveRange(30, myFileHistoryList.Count - 30);
+            }
             return myFileHistoryList;
         }
     }
